[Diffusion] Reuse shared AlignedVector and tidy jit_kernel/diffusion
Reuse the shared device::AlignedVector component across the KDA-Pilot
diffusion CUDA kernels instead of each kernel hand-rolling its own
vectorization struct, and apply a few small, verified cleanups elsewhere
in jit_kernel/diffusion.

Native CUDA (.cuh):
- residual_gate_add: replace `union Vec16<T>` with device::AlignedVector
- causal_conv3d_cat_pad: replace `union Pack` with device::AlignedVector
- timestep_embedding: replace raw float4 with AlignedVector<float, 4>,
  add `#pragma once` and a named namespace to match its siblings (and
  the matching wrapper symbol in timestep_embedding.py)

cutedsl / triton cleanups:
- hoist the byte-identical to_cute_arg / to_fake_cute_args helpers into
  cutedsl/utils.py and import them in both kernel files
- fix a copy-paste guard in norm_tanh_mul_add_norm_scale (isinstance(src)
  twice -> src/dst)
- drop the dead _precompute_inv_rms (superseded by fused_qk_inv_rms)
- drop the unused `rows` param from the scale_shift 4D kernel
- drop unused B,S args from validate_weight_bias (only D is used)
- remove stale commented-out code / fix a stale docstring module name

No behavior change. Validated on B200: existing diffusion unit tests pass
(residual_gate_add, causal_conv3d_cat_pad, timestep_embedding,
fused_norm_scale_shift, qwen_image_modulation), a torch-reference smoke
over every touched kernel matches, and residual_gate_add /
causal_conv3d_cat_pad before/after benchmarks are within run-to-run noise.
