+++ python/sglang/jit_kernel/diffusion/cutedsl/norm_tanh_mul_add_norm_scale.py
@@ -10,48 +10,13 @@ from sglang.jit_kernel.diffusion.cutedsl.common.norm_fusion import (
     broadcast_tensor_for_bsfd,
     tensor_slice_for_bsfd,
 )
-from sglang.jit_kernel.diffusion.cutedsl.utils import TORCH_TO_CUTE_DTYPE, WARP_SIZE
+from sglang.jit_kernel.diffusion.cutedsl.utils import (
+    WARP_SIZE,
+    to_cute_arg,
+    to_fake_cute_args,
+)
 
 _COMPILE_CACHE = {}
-
-
-def to_cute_arg(
-    t,
-    *,
-    assume_aligned: Optional[int] = 32,
-    use_32bit_stride: bool = False,
-    enable_tvm_ffi: bool = True,
-):
-    """
-    Convert a Python value into a CuTeDSL value.
-    """
-    if isinstance(t, torch.Tensor):
-        return cute.runtime.from_dlpack(
-            t,
-            assumed_align=assume_aligned,
-            use_32bit_stride=use_32bit_stride,
-            enable_tvm_ffi=enable_tvm_ffi,
-        )
-    if isinstance(t, int):
-        return cutlass.Int32(t)
-    if isinstance(t, float):
-        return cutlass.Float32(t)
-    return t
-
-
-def to_fake_cute_args(t: torch.Tensor):
-    if isinstance(t, torch.Tensor):
-        # Only keep the last dim as compile-time value to maximum compiled kernel reuse
-        # e.g. (1,2,1536):(3027,1536,1) -> (?,?,1536):(?,?,1)
-        D = t.shape[-1]
-        dtype = TORCH_TO_CUTE_DTYPE[t.dtype]
-        shape = (*(cute.sym_int() for _ in range(t.ndim - 1)), D)
-        stride = (*(cute.sym_int(divisibility=D) for _ in range(t.ndim - 1)), 1)
-        fake_t = cute.runtime.make_fake_tensor(
-            dtype, shape, stride, memspace=cute.AddressSpace.gmem, assumed_align=32
-        )
-        return fake_t
-    return to_cute_arg(t)
 
 
 class NormTanhMulAddNormScale:
@@ -166,7 +131,7 @@ class NormTanhMulAddNormScale:
         @cute.jit
         def copy_if(src, dst):
             if cutlass.const_expr(
-                isinstance(src, cute.Tensor) and isinstance(src, cute.Tensor)
+                isinstance(src, cute.Tensor) and isinstance(dst, cute.Tensor)
             ):
                 cute.autovec_copy(src, dst)  # LDG.128
