+++ python/sglang/jit_kernel/diffusion/cutedsl/scale_residual_norm_scale_shift.py
@@ -10,48 +10,12 @@ from sglang.jit_kernel.diffusion.cutedsl.common.norm_fusion import (
     broadcast_tensor_for_bsfd,
     tensor_slice_for_bsfd,
 )
-from sglang.jit_kernel.diffusion.cutedsl.utils import TORCH_TO_CUTE_DTYPE, WARP_SIZE
+from sglang.jit_kernel.diffusion.cutedsl.utils import (
+    WARP_SIZE,
+    to_fake_cute_args,
+)
 
 _COMPILE_CACHE = {}
-
-
-def to_cute_arg(
-    t,
-    *,
-    assume_aligned: Optional[int] = 32,
-    use_32bit_stride: bool = False,
-    enable_tvm_ffi: bool = True,
-):
-    """
-    Convert a Python value into a CuTeDSL value.
-    """
-    if isinstance(t, torch.Tensor):
-        return cute.runtime.from_dlpack(
-            t,
-            assumed_align=assume_aligned,
-            use_32bit_stride=use_32bit_stride,
-            enable_tvm_ffi=enable_tvm_ffi,
-        )
-    if isinstance(t, int):
-        return cutlass.Int32(t)
-    if isinstance(t, float):
-        return cutlass.Float32(t)
-    return t
-
-
-def to_fake_cute_args(t: torch.Tensor):
-    if isinstance(t, torch.Tensor):
-        # Only keep the last dim as compile-time value to maximum compiled kernel reuse
-        # e.g. (1,2,1536):(3027,1536,1) -> (?,?,1536):(?,?,1)
-        D = t.shape[-1]
-        dtype = TORCH_TO_CUTE_DTYPE[t.dtype]
-        shape = (*(cute.sym_int() for _ in range(t.ndim - 1)), D)
-        stride = (*(cute.sym_int(divisibility=D) for _ in range(t.ndim - 1)), 1)
-        fake_t = cute.runtime.make_fake_tensor(
-            dtype, shape, stride, memspace=cute.AddressSpace.gmem, assumed_align=32
-        )
-        return fake_t
-    return to_cute_arg(t)
 
 
 class ScaleResidualNormScaleShift:
@@ -234,7 +198,7 @@ def validate_x(t: torch.Tensor, B: int, S: int, D: int):
         raise ValueError(f"Validate failed: not contiguous on dim D.")
 
 
-def validate_weight_bias(t: Optional[torch.Tensor], B: int, S: int, D: int):
+def validate_weight_bias(t: Optional[torch.Tensor], D: int):
     if t is None:
         return
     if t.dtype not in (torch.float16, torch.bfloat16, torch.float32):
@@ -312,8 +276,8 @@ def fused_norm_scale_shift(
     # Tensor Validation
     BSD = x.shape
     validate_x(x, *BSD)
-    validate_weight_bias(weight, *BSD)
-    validate_weight_bias(bias, *BSD)
+    validate_weight_bias(weight, BSD[-1])
+    validate_weight_bias(bias, BSD[-1])
     validate_scale_shift(scale, *BSD)
     validate_scale_shift(shift, *BSD)
 
@@ -399,14 +363,13 @@ def fused_scale_residual_norm_scale_shift(
     validate_x(x, *BSD)
     validate_x(residual, *BSD)
     validate_gate(gate, *BSD)
-    validate_weight_bias(weight, *BSD)
-    validate_weight_bias(bias, *BSD)
+    validate_weight_bias(weight, BSD[-1])
+    validate_weight_bias(bias, BSD[-1])
     validate_scale_shift(scale, *BSD)
     validate_scale_shift(shift, *BSD)
     if norm_type == "layer" or norm_type == "rms":
         stream = cuda.CUstream(torch.cuda.current_stream().cuda_stream)
 
-        # if norm_type == "layer" or norm_type == "rms":
         D = x.shape[-1]
         if D % 256 != 0 or D > 8192:
             raise ValueError(
